requirements.txt
appdirs==1.4.4
asgiref==3.2.7
Babel==2.8.0
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
django-import-export==2.1.0
django-phonenumber-field==4.0.0
django-six==1.0.4
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
html5lib==1.0.1
importlib-metadata==1.6.0
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
phonenumbers==8.12.4
phonenumberslite==8.12.4
Pillow==7.1.2
psycopg2==2.8.5
PyPDF2==1.26.0
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.42
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
virtualenv==20.0.21
webencodings==0.5.1
whitenoise==5.1.0
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
